Modularizing Spins page to work more like the SPA of the main site.
