+++ src/content/SpinsIntro.md
@@ -1,4 +1,4 @@
-### What are Spins ?
+# What are Spins ?
 
 Spins are alternate versions of XeroLinux, using other Desktop Environments, for those not into KDE but enjoy my tweaks, they are created & optimized for your consumption. But unlike the Flagship, they'll not change or be modified beyond updating the packages underneath. They'll remain as they are, until their associated Destop Environment gets a major update, for example, the next XeroG release will be once Gnome 45 releases, XFCE 4.19 & so on. Also, spins will receive limited support (Forums, Discord) since I, myself do not use them daily...
