+++ src/content/SpinsOutro.md
@@ -1,11 +1,12 @@
-### Why It's Behind a Donation Wall :
+# Why It's Behind a Donation Wall :
  
 Well, as you all know by now, my current situation, in Lebanon, is not great, with up to 4h or Government provided electricity, I am limited on time & it's much harder to maintain such things. With no other income, to rent Generator power, I rely on your generous contributions. So, as a result I had no choice but to charge for my time.. See video below.
 
 <center>
 <iframe
-    width="640"
-    height="480"
+    id="lebanon-video"
+    width="1088"
+    height="612"
     src="https://www.youtube.com/embed/cUxDtW8Zddc"
     frameborder="0"
     allow="autoplay; encrypted-media"
