+++ src/index.css
@@ -136,3 +136,13 @@ img {
 .carousel {
   text-align: center;
 }
+
+.carousel-slider {
+  border: var(--border) solid rgb(255, 255, 255);
+  border-radius: var(--border-radius);
+}
+
+#lebanon-video {
+  border: var(--border) solid rgb(var(--fg));
+  border-radius: var(--border-radius);
+}
